Reupload Player Movement script
This script adds code to the Jump() function, which calls checkIfGrounded() to help detect if the player should be able to jump/not jump. 
Infinite jumping should not be allowed, and this should enforce that.
Be careful not to copy this exactly onto our new projects because it has different dependencies than our newer movement script called "Player_Controller.cs"

It depends on the Transform and LayerMask components, which should be added to the Player sprite in the correct way.
Will needed tested in Unity again.