Tile palette with basic ruleset
This zip contains a set of prefabs for a tile palette that can be used to paint a level with tilemaps.
They are saved as prefabs, so it should be mostly functional if you put them inside your own unity project.
You'll need the metadata file along with it.

Additionally, I added a short word document that describes how you can replicate the process I used to create/edit this tile palette. 
Please read this because you need to download the "2D Tile Map Extras" from Unity package manager before this works.